requirements.txt
kopf==1.37.1
pydantic==2.5.2
pydantic-settings==2.1.0
pykube-ng==23.6.0
